refact: move :composite:text to :composite:timeline
