--- composite/text/build.gradle.kts
@@ -1,37 +1,0 @@
-/*
- * Copyright © 2024 Orca
- *
- * This program is free software: you can redistribute it and/or modify it under the terms of the
- * GNU General Public License as published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
- * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with this program. If
- * not, see https://www.gnu.org/licenses.
- */
-
-plugins {
-  alias(libs.plugins.android.library)
-  alias(libs.plugins.kotlin.android)
-}
-
-android.defaultConfig.testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-
-dependencies {
-  androidTestImplementation(libs.android.test.core)
-  androidTestImplementation(libs.android.test.runner)
-  androidTestImplementation(libs.assertk)
-
-  api(project(":platform:autos"))
-  api(project(":std:styled-string"))
-  api(libs.android.compose.foundation)
-
-  implementation(libs.android.core)
-  implementation(libs.jsoup)
-
-  testImplementation(libs.assertk)
-  testImplementation(libs.kotlin.test)
-}
