+++ composite/timeline-test/build.gradle.kts
@@ -28,7 +28,6 @@ android {
 }
 
 dependencies {
-  androidTestImplementation(project(":platform:autos"))
   androidTestImplementation(libs.android.compose.ui.test.junit)
   androidTestImplementation(libs.android.compose.ui.test.manifest)
   androidTestImplementation(libs.android.test.core)
