+++ composite/timeline/src/androidTest/java/com/jeanbarrossilva/orca/composite/timeline/text/StyledStringExtensionsTests.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2023 Orca
+ * Copyright © 2023-2024 Orca
  *
  * This program is free software: you can redistribute it and/or modify it under the terms of the
  * GNU General Public License as published by the Free Software Foundation, either version 3 of the
@@ -13,27 +13,16 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package com.jeanbarrossilva.orca.composite.text.spanned
+package com.jeanbarrossilva.orca.composite.timeline.text
 
-import android.graphics.Typeface
-import android.text.Html
-import android.text.style.StyleSpan
 import assertk.assertThat
-import assertk.assertions.containsExactly
+import assertk.assertions.isEqualTo
+import com.jeanbarrossilva.orca.std.styledstring.StyledString
 import org.junit.Test
 
-internal class SpannedExtensionsTests {
+internal class StyledStringExtensionsTests {
   @Test
-  fun partitions() {
-    assertThat(
-        Html.fromHtml("<p><b><i>Hello</i></b>, <i>world</i>!</p>", Html.FROM_HTML_MODE_COMPACT)
-          .parts
-      )
-      .containsExactly(
-        Part(0..4).span(StyleSpan(Typeface.ITALIC)),
-        Part(0..4).span(StyleSpan(Typeface.BOLD, fontWeightAdjustment = 0)),
-        Part(5..6),
-        Part(7..11).span(StyleSpan(Typeface.ITALIC))
-      )
+  fun breaksLineTwiceBetweenParagraphsWhenConvertingHtmlToStyledString() {
+    assertThat(StyledString.fromHtml("<p>👔</p><p>🥾</p>")).isEqualTo(StyledString("👔\n\n🥾"))
   }
 }
