+++ composite/timeline/src/androidTest/java/com/jeanbarrossilva/orca/composite/timeline/text/spanned/SpannedExtensionsTests.kt
@@ -13,16 +13,27 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package com.jeanbarrossilva.orca.composite.text
+package com.jeanbarrossilva.orca.composite.timeline.text.spanned
 
+import android.graphics.Typeface
+import android.text.Html
+import android.text.style.StyleSpan
 import assertk.assertThat
-import assertk.assertions.isEqualTo
-import com.jeanbarrossilva.orca.std.styledstring.StyledString
+import assertk.assertions.containsExactly
 import org.junit.Test
 
-internal class StyledStringExtensionsTests {
+internal class SpannedExtensionsTests {
   @Test
-  fun breaksLineTwiceBetweenParagraphsWhenConvertingHtmlToStyledString() {
-    assertThat(StyledString.fromHtml("<p>👔</p><p>🥾</p>")).isEqualTo(StyledString("👔\n\n🥾"))
+  fun partitions() {
+    assertThat(
+        Html.fromHtml("<p><b><i>Hello</i></b>, <i>world</i>!</p>", Html.FROM_HTML_MODE_COMPACT)
+          .parts
+      )
+      .containsExactly(
+        Part(0..4).span(StyleSpan(Typeface.ITALIC)),
+        Part(0..4).span(StyleSpan(Typeface.BOLD, fontWeightAdjustment = 0)),
+        Part(5..6),
+        Part(7..11).span(StyleSpan(Typeface.ITALIC))
+      )
   }
 }
