+++ composite/timeline/src/main/java/com/jeanbarrossilva/orca/composite/timeline/post/Post.extensions.kt
@@ -17,11 +17,11 @@ package com.jeanbarrossilva.orca.composite.timeline.post
 
 import androidx.compose.runtime.Composable
 import com.jeanbarrossilva.orca.autos.colors.Colors
-import com.jeanbarrossilva.orca.composite.text.toAnnotatedString
 import com.jeanbarrossilva.orca.composite.timeline.post.figure.Figure
 import com.jeanbarrossilva.orca.composite.timeline.post.figure.gallery.disposition.Disposition
 import com.jeanbarrossilva.orca.composite.timeline.post.figure.gallery.thumbnail.Thumbnail
 import com.jeanbarrossilva.orca.composite.timeline.stat.details.asStatsDetails
+import com.jeanbarrossilva.orca.composite.timeline.text.toAnnotatedString
 import com.jeanbarrossilva.orca.core.feed.profile.post.Post
 import com.jeanbarrossilva.orca.core.feed.profile.post.repost.Repost
 import com.jeanbarrossilva.orca.platform.autos.theme.AutosTheme
