+++ composite/timeline/src/main/java/com/jeanbarrossilva/orca/composite/timeline/text/spanned/Spanned.extensions.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2023 Orca
+ * Copyright © 2023-2024 Orca
  *
  * This program is free software: you can redistribute it and/or modify it under the terms of the
  * GNU General Public License as published by the Free Software Foundation, either version 3 of the
@@ -13,7 +13,7 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package com.jeanbarrossilva.orca.composite.text.spanned
+package com.jeanbarrossilva.orca.composite.timeline.text.spanned
 
 import android.text.Spanned
 import androidx.core.text.getSpans
