+++ core/mastodon/build.gradle.kts
@@ -42,7 +42,6 @@ android {
 
 dependencies {
   androidTestImplementation(project(":platform:autos-test"))
-  androidTestImplementation(project(":platform:core"))
   androidTestImplementation(project(":platform:intents"))
   androidTestImplementation(project(":platform:testing"))
   androidTestImplementation(project(":std:injector-test"))
@@ -56,7 +55,7 @@ dependencies {
 
   ksp(project(":std:injector-processor"))
 
-  implementation(project(":composite:text"))
+  implementation(project(":composite:timeline"))
   implementation(project(":core:sample"))
   implementation(project(":ext:coroutines"))
   implementation(project(":platform:autos"))
