+++ feature/profile-details/src/main/java/com/jeanbarrossilva/orca/feature/profiledetails/conversion/converter/DefaultProfileConverter.kt
@@ -16,7 +16,7 @@
 package com.jeanbarrossilva.orca.feature.profiledetails.conversion.converter
 
 import com.jeanbarrossilva.orca.autos.colors.Colors
-import com.jeanbarrossilva.orca.composite.text.toAnnotatedString
+import com.jeanbarrossilva.orca.composite.timeline.text.toAnnotatedString
 import com.jeanbarrossilva.orca.core.feed.profile.Profile
 import com.jeanbarrossilva.orca.feature.profiledetails.ProfileDetails
 import com.jeanbarrossilva.orca.feature.profiledetails.conversion.ProfileConverter
