+++ feature/profile-details/src/test/java/com/jeanbarrossilva/orca/feature/profiledetails/test/ProfileDetails.Editable.extensions.kt
@@ -16,7 +16,7 @@
 package com.jeanbarrossilva.orca.feature.profiledetails.test
 
 import com.jeanbarrossilva.orca.autos.colors.Colors
-import com.jeanbarrossilva.orca.composite.text.toAnnotatedString
+import com.jeanbarrossilva.orca.composite.timeline.text.toAnnotatedString
 import com.jeanbarrossilva.orca.core.feed.profile.type.editable.EditableProfile
 import com.jeanbarrossilva.orca.core.sample.test.feed.profile.type.sample
 import com.jeanbarrossilva.orca.feature.profiledetails.ProfileDetails
