+++ feature/search/build.gradle.kts
@@ -31,7 +31,6 @@ dependencies {
 
   implementation(project(":composite:timeline"))
   implementation(project(":core:sample"))
-  implementation(project(":platform:autos"))
   implementation(project(":platform:core"))
   implementation(project(":platform:starter"))
   implementation(project(":platform:ui"))
